UPD: Person class

+++ src/warfake/war/race/Person.java
@@ -1,19 +1,19 @@
 package warfake.war.race;
 
 public abstract class Person {
-	private int HP = 100;
-	private static final int MAX_HP = 100;
+	private float HP = 100;
+	private static final float MAX_HP = 100;
 	private int accuracy = 100;
 
-	public int getHealth() {
+	public float getHealth() {
 		return HP;
 	}
 
-	public int getMaxHealth() {
+	public float getMaxHealth() {
 		return MAX_HP;
 	}
 
-	public void setHealth(int health) {
+	public void setHealth(float health) {
 		this.HP = health;
 	}
